Remove deplicate function
